fixed build errors on prospero build

+++ Include/BasicDataTypes/Bezier.h
@@ -50,7 +50,7 @@ namespace CE
 	{
 		int num{};
 
-		for (;num < std::tuple_size_v<ValueStorage>; num++)
+		for (;static_cast<size_t>(num) < std::tuple_size_v<ValueStorage>; num++)
 		{
 			if (key[num * 2] < 0.0f
 				|| key[num * 2] > 1.0f)
